Moved PresetEditor into Dialogs subdir
